Added a dummy code of conduct
Note that the Github web interface only allows one to select between two different Codes of Conducts for some reason, neither of which are the ones we agreed with the CCC during project adoption, but creating a Code of Conduct file through the web interface is the only way to avoid it thinking that there is no such file!  As a result, I'm adding this file which will be updated with the real Code of Conduct once merged.